requirements.txt
numpy==2.2.5
psycopg2-binary==2.9.10
langchain==0.3.24
python-dotenv==1.0.1
openai